LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-multicol-002-expected.txt










































PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="35"><span></span></div>
offsetLeft expected 35 but got 10
FAIL .target > * 4 assert_equals:
<div class="multicol" data-offset-x="10">
    <div style="font-size: 10px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 30px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 20px;">
      <span></span><br><span></span><br><span></span>
    </div>
  </div>
offsetLeft expected 10 but got 140
FAIL .target > * 5 assert_equals:
<div data-offset-x="140"><span></span></div>
offsetLeft expected 140 but got 145
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="35"><span></span></div>
offsetLeft expected 35 but got 120
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
